Generation de vis modifiee

+++ cylinder.h
@@ -9,9 +9,9 @@ public:
 
     float radius;
     float length;
-    float precision;
+    int precision;
 
-    Cylinder(float radius, float length, float precision, V3 center = V3(0,0,0), V3 rotation = V3(0,0,0));
+    Cylinder(float radius, float length, int precision, V3 center = V3(0,0,0), V3 rotation = V3(0,0,0));
     Cylinder();
     void create_vertices() override;
     void create_mesh() override;
